requirements.txt
langgraph
langchain-anthropic
streamlit
python-dotenv
ddgs
anthropic